pyproject.toml
[tool.poetry]
name = "hk69"
version = "0.1.0"
description = ""
authors = ["zeffo <phatakaman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = {git = "https://github.com/Rapptz/discord.py"}
PyYAML = "^6.0"
pydantic = "^1.9.0"
psutil = "^5.9.0"
jishaku = "^2.5.0"
yt-dlp = "^2022.4.8"
PyNaCl = "^1.5.0"
prisma = "^0.6.5"
colorthief = "^0.2.1"
numpy = "^1.23.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
